fix(pt): change `getCheckResultString`、`getPositionChangeString` to shared utils
